[Fix] #240 StudyData 모델 변경
