+++ app/src/main/java/kr/co/lion/modigm/db/write/WriteStudyDataSource.kt
@@ -1,12 +1,12 @@
 package kr.co.lion.modigm.db.write
 
 import android.util.Log
-import kr.co.lion.modigm.model.StudyData
+import kr.co.lion.modigm.model.SqlStudyData
 
 class WriteStudyDataSource {
 
     // 사용자 정보를 저장한다.
-    suspend fun uploadStudyData(study: StudyData):Int?{
+    suspend fun uploadStudyData(study: SqlStudyData):Int?{
         return try {
             val db = WriteStudyDao()
             db.insertStudyData(study)
